feat: switch return type\n\nBREAKING CHANGE: parse returns an object instead of an array
