+++ src/parsers/index.js
@@ -1,0 +1,9 @@
+import date from './date.js'
+import time from './time.js'
+import duration from './duration.js'
+import list from './list.js'
+
+export const parseDate = date
+export const parseTime = time
+export const parseDuration = duration
+export const parseList = list
