+++ src/parsers/list.js
@@ -1,0 +1,29 @@
+export default function parseList(list) {
+  return list.children
+    .map((item) => {
+      const listItem = {}
+      if (item.type === 'list') {
+        return parseList(list)
+      } else if (item.type === 'listItem') {
+        listItem.checked = item.checked
+        return item.children
+          .map((child) => {
+            if (child.type === 'paragraph') {
+              listItem.text = child.children
+                .map((c) => {
+                  if (c.type === 'link') {
+                    return c.children[0].value
+                  } else {
+                    return c.value
+                  }
+                })
+                .filter((x) => !!x)
+                .join('')
+              return listItem
+            }
+          })
+          .filter((x) => !!x)
+      }
+    })
+    .filter((x) => !!x)
+}
