+++ test/parse-issue.test.js
@@ -8,32 +8,32 @@ import { join } from 'path'
 const test = t.test
 
 test('parse(md) should parse GitHub Issue Form data into useful, structured data', async (t) => {
-  const expected = [
-    {
-      id: 'event-description',
+  const expected = {
+    'event-description': {
+      order: 0,
       title: 'Event Description',
       text: "Let's meet for coffee and chat about tech, coding, Cyprus and the newly formed\nCDC (Cyprus Developer Community)."
     },
-    {
-      id: 'location',
+    location: {
+      order: 1,
       title: 'Location',
       text: '[Cafe Nero Finikoudes, Larnaka](https://goo.gl/maps/Bzjxdeat3BSdsUSVA)'
     },
-    {
-      id: 'date',
+    date: {
+      order: 2,
       title: 'Date',
       text: '11.03.2022',
       date: '2022-03-11'
     },
-    { id: 'time', title: 'Time', text: '16:00', time: '16:00' },
-    {
-      id: 'duration',
+    time: { order: 3, title: 'Time', text: '16:00', time: '16:00' },
+    duration: {
+      order: 4,
       title: 'Duration',
       text: '2h',
       duration: { hours: 2, minutes: 0 }
     },
-    {
-      id: 'list-item-checked',
+    'list-item-checked': {
+      order: 5,
       title: 'List Item Checked',
       list: [
         {
@@ -43,8 +43,8 @@ test('parse(md) should parse GitHub Issue Form data into useful, structured data
       ],
       text: "*   [x] I agree to follow this project's\n    [Code of Conduct](https://berlincodeofconduct.org)"
     },
-    {
-      id: 'list-item-unchecked',
+    'list-item-unchecked': {
+      order: 6,
       title: 'List Item Unchecked',
       list: [
         {
@@ -54,8 +54,8 @@ test('parse(md) should parse GitHub Issue Form data into useful, structured data
       ],
       text: "*   [ ] I agree to follow this project's\n    [Code of Conduct](https://berlincodeofconduct.org)"
     },
-    {
-      id: 'mixed-task-list',
+    'mixed-task-list': {
+      order: 7,
       title: 'Mixed Task List',
       list: [
         { checked: true, text: 'checked' },
@@ -66,8 +66,8 @@ test('parse(md) should parse GitHub Issue Form data into useful, structured data
       ],
       text: '*   [x] checked\n*   [ ] unchecked\n*   [x] checked 2\n*   [x] checked 3\n*   [ ] unchecked 2'
     },
-    {
-      id: 'complex-list',
+    'complex-list': {
+      order: 8,
       title: 'Complex List',
       list: [
         { checked: null, text: 'one' },
@@ -75,17 +75,17 @@ test('parse(md) should parse GitHub Issue Form data into useful, structured data
       ],
       text: '*   one\n*   two\n    *   three\n    *   four\n        1.  five\n        2.  six'
     },
-    {
-      id: 'repositories',
+    repositories: {
+      order: 9,
       title: 'Repositories',
       text: '```csv\nhttps://example.com/repository-1\nhttps://example.com/repository-2\n```'
     },
-    {
-      id: 'visibility',
+    visibility: {
+      order: 10,
       title: 'Visibility',
       text: 'Internal'
     }
-  ]
+  }
 
   const md = await readFile(
     join(process.cwd(), 'test', 'test-issue-1.md'),
@@ -97,7 +97,7 @@ test('parse(md) should parse GitHub Issue Form data into useful, structured data
 })
 
 test('parse(md) return nothing', async (t) => {
-  const expected = []
+  const expected = {}
 
   const md = await readFile(
     join(process.cwd(), 'test', 'test-issue-2.md'),
